+++ hackage-security-http-client/src/Hackage/Security/Client/Repository/HttpLib/HttpClient.hs
@@ -21,7 +21,6 @@ import qualified Network.HTTP.Types           as HttpClient
 import Hackage.Security.Client hiding (Header)
 import Hackage.Security.Client.Repository.HttpLib
 import Hackage.Security.Util.Checked
-import qualified Hackage.Security.Util.Lens as Lens
 
 {-------------------------------------------------------------------------------
   Top-level API
@@ -161,7 +160,14 @@ setRequestHeaders opts req = req {
     finalizeHeader (name, strs) = [(name, BS.intercalate ", " (reverse strs))]
 
     insert :: Eq a => a -> [b] -> [(a, [b])] -> [(a, [b])]
-    insert x y = Lens.modify (Lens.lookupM x) (++ y)
+    insert x y = modifyAssocList x (++ y)
+
+    -- modify the first maching element
+    modifyAssocList :: Eq a => a -> (b -> b) -> [(a, b)] -> [(a, b)]
+    modifyAssocList a f = go where
+        go []                           = []
+        go (p@(a', b) : xs) | a == a'   = (a', f b) : xs
+                            | otherwise = p         : go xs
 
 -- | Extract the response headers
 getResponseHeaders :: HttpClient.Response a -> [HttpResponseHeader]