+++ hackage-security/ChangeLog.md
@@ -9,6 +9,7 @@ See also http://pvp.haskell.org/faq
   `ReportSchemaErrors(expected)` instead of via `Monad(fail)`
 * Add support for GHC 8.8 / base-4.13
 * Use `lukko` for file-locking
+* Don't expose `Hackage.Security.Util.Lens` module
 
 0.5.3.0
 -------